통합테스트까지 완료, 단위테스트 할게있남..

+++ src/main/java/com/homework/passionfactory/domain/todo/dao/TodoRepository.java
@@ -3,5 +3,5 @@ package com.homework.passionfactory.domain.todo.dao;
 import com.homework.passionfactory.domain.todo.domain.Todo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TodoRepository extends JpaRepository<Todo,Long> {
+public interface TodoRepository extends JpaRepository<Todo,Integer> {
 }
